refactor: 달력 백엔드 파일을 core 패키지로 옮기고, SwipeableCalendar와 Calendar를 분리
