+++ app/src/test/java/com/practice/hanbitlunch/calendar/CalendarPageTest.kt
@@ -1,6 +1,7 @@
 package com.practice.hanbitlunch.calendar
 
 import com.hsk.ktx.date.Date
+import com.practice.hanbitlunch.calendar.core.CalendarPage
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
